chore: updates as per PR feedback - beta
