+++ packages/wallets/wallet-core/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './tx.js'
-export * from './MsgBroadcaster.js'
