+++ packages/wallets/wallet-core/src/utils/tx.ts
@@ -3,6 +3,8 @@ import {
   ChainCosmosErrorCode,
   TransactionChainErrorModule,
 } from '@injectivelabs/exceptions'
+import { NetworkEndpoints } from '@injectivelabs/networks'
+import { TxResponse, ChainGrpcTendermintApi } from '@injectivelabs/sdk-ts'
 
 export const checkIfTxRunOutOfGas = (e: unknown) => {
   return (
@@ -11,4 +13,28 @@ export const checkIfTxRunOutOfGas = (e: unknown) => {
     e.contextModule === TransactionChainErrorModule.CosmosSdk &&
     e.originalMessage.includes('out of gas')
   )
+}
+
+export const getTransactionTimeElapsed = async ({
+  endpoints,
+  txResponse,
+}: {
+  txResponse: TxResponse
+  endpoints: NetworkEndpoints
+}) => {
+  const endTimeTx = txResponse.timestamp
+    ? new Date(txResponse.timestamp).getTime()
+    : 0
+
+  const txBlock = await new ChainGrpcTendermintApi(endpoints.grpc).fetchBlock(
+    txResponse.height,
+  )
+
+  const endTimeTxBlock = txBlock?.header?.time
+    ? new Date(txBlock?.header?.time).getTime()
+    : 0
+
+  const timeElapsed = endTimeTxBlock - endTimeTx
+
+  return timeElapsed
 }
